public/skills/bioengine/bioengine_cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bioengine"
version = "0.6.15"
description = "BioEngine CLI — deploy, manage, and call BioEngine AI model services from the command line"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "click>=8.1",
    "hypha-rpc>=0.21.11",
    "httpx>=0.28",
    "numpy>=1.24",
    "tifffile>=2023.1.1",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
bioengine = "bioengine_cli.cli:main"

[tool.setuptools.packages.find]
where = [".."]
include = ["bioengine_cli*"]
